+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file. This projec
 
 ## [Unreleased](https://github.com/Shatur95/akd/tree/HEAD)
 
-[Full Changelog](https://github.com/Shatur95/akd/compare/2.0.0...HEAD)
+[Full Changelog](https://github.com/Shatur95/akd/compare/2.1.0...HEAD)
+
+_No unreleased changes yet_
+
+## [2.1.0](https://github.com/Shatur95/akd/tree/2.1.0) (2020-04-25)
+
+[Full Changelog](https://github.com/Shatur95/akd/compare/2.1.0...2.0.0)
 
 **Added**